fix styles and biginteger

+++ database/migrations/2023_09_01_221240_create_creditos_table.php
@@ -16,7 +16,7 @@ class CreateCreditosTable extends Migration
         Schema::create('creditos', function (Blueprint $table) {
             $table->id();
             $table->text('nombre')->nullable();
-            $table->decimal('cantidad', 9,2);
+            $table->bigInteger('cantidad');
             $table->integer('cuotas');
             $table->decimal('interes',4,2);
             $table->unsignedBigInteger('user_id');
